feat: use concept IDs in grouping first (#575)
This looks like a lot of code changes but it's actually a fairly simple fix (that was made more complicated by aforementioned spaghettiness). Basically, when trying to group a drug or gene, we now do three steps

    Collect claim name/aliases that have nomenclature values taken from a select set of identifier spaces (e.g., for genes, collect all name/aliases that are an HGNC ID, an NCBI ID, or an Ensembl ID). Sort them by an order of preference (for genes, HGNC > NCBI > Ensembl). Take the first match. This is the new part.
    If that fails, and if the claim name wasn't one of the previously selected IDs, run it through the normalizer, and put it in that group if it matches.
    Failing that, do the previous alias-based "voting" that we'd been falling back on previously. Two changes here
        Since we're pulling out a selection of xref-like aliases above, they aren't included here
        I got rid of the "tiebreaker" logic at the very bottom (where, if you have a tied # of normalizer "alias" matches, you take the one that's alphabetically earlier). When I looked at this a few months ago, it happened a total of 1 time across all drug and gene groups, but it seems like a bad idea (better to refuse to group when we're unsure).

Previously, we searched on the claim name first, but that can be an ambiguous symbol (a drug named "ADM" can resolve in multiple ways). By first searching on a selection of high-quality identifiers, we get a more reliable grouping outcome. I intend to do a little more testing, but this should close #567.